Fix tile spacing issue
When calculating tile size, some rounding happens because pixels are
discrete values. Because of this, providing a hard-coded spacing between
tiles can cause the final column to have a slightly smaller width than
the rest of the columns.

To fix this, configure the rows and columns to render the tiles with
space between entries, rather than hard-coding the spacing value.