Add DragContext and implement basic drag-and-drop
Introduce the `DragContext` class, which holds the state of registered
draggable and droppable UI elements ("drag targets" and "drop targets"),
such as dragged/hovered state and offset amount. Importantly, callbacks
can be configured to fire when "drop" events happen in order to
implement drag-and-drop behavior.

Registration of a drag target requires simply wrapping the desired
element in the `DragTarget` composable, which will create and cache the
appropriate state in the context. When parameters to `DragTarget`
change, or if the drag target leaves the composition entirely, then
automatic cleanup ensures that the associated state is removed from the
context. Registration of drop targets is done similarly with the
`DropTarget` wrapping composable.

Because `DragTarget` and `DropTarget` must be called within the context
of a `DragContext`, the `withDragContext` convenience function is
provided so that users can wrap appropriate UI elements in the necessary
drag/drop target wrappers. Though not necessary, creating a
`CompositionLocal` of the drag context may be helpful in order to avoid
having to pass down the context to arbitrarily deep children in the
composition tree.

An initial implementation of drag-and-drop has been added for placing
tiles onto the board. Actual drop behavior (e.g. checking validity,
calculating points, etc) will happen in a future commit.