Make isDragging state consistent in DragContext
Previously, `isDragging` was intentionally kept as `true` upon pointer
release if the drag target was over a drop target.

This was done for two reasons:
1) We want to maintain the offset of the target upon release (e.g. for
cases where the target may be dragged again in the future)
2) Applying the current fix of accounting for scaling caused another
issue that has since been fixed

Applying this fix makes it easier to reason about drag behavior - i.e.
`isDragging` is true only while a user is actually dragging the target.