Fix bug in snap-to-target logic
There was a bug in the setting of the snap offset after a drag target
has been dropped. The issue was that `dragOffsetState` always contains a
value that has been scaled to reflect the provided `onDragScaleX` and
`onDragScaleY`, but the snap offset (i.e. the amount that we need to
shift the drag target by in order to snap to its drop target) did not
include this scaling factor. Therefore, in the current use case where
the target is scaled down while dragging, the `dragOffsetState` was not
being updated by enough pixels.

Interestingly, this manifested itself as a "bug as a feature" where,
because the offset was never actually as large as it needed to be in
order to snap in one movement, the drag target was "animated" to its
final snap location (e.g. if the scale factor was 50%, it would get 50%
of the way to the final destination repeatedly until it actually reached
the final position). This provided a smooth UX, and perhaps this can be
introduced in the future as an intentional feature/UX improvement.