Implement rearranging tiles in the tiles section
Originally, I had experimented with implementing this with the
DragContext API, but it turns out that the API was simultaneously not
flexible enough in ways that I needed and had more functionality than
was necessary for this simpler use case (e.g. no data to drop).

In the end, the implementation was a purely graphical manipulation of
offsets; in other words, no state is changed when rearranging tiles.

Side note: Because of the drag-and-drop nature of this game, we have to
be careful about which graphics layer that compose is drawing into. One
such example of this gotcha is that I wanted to clip the tiles in the
tile container with a RoundedCornerShape, but soon realized that doing
so causes the entire composable to be drawn in a different graphics
layer. Not a big deal, but this imposes some UI/stylistic limitations.