Make grid cell responsible for displaying placed tile
Due to the way that snap-to-target is implemented in DragContext, a
placed tile will visibly jitter if its parent were to change in position
while shuffling or rearranging the tiles. This is not a bug; it is due
to the constant re-cecalculation that is occurring in order to remain
snapped to the correct grid cell location.

To avoid this jitter, we now hide the tile composable when in a dropped
state and make it the responsibility of the grid cell to display the
tile information, which it already has access to.