Require nullability to be explicit in DragContext
We were already accepting a type parameter in DragContext for the data
to be dropped from the drag target into the drop target. However, in
reality, we were allowing for a DragContext<T> to handle data of type T
*or* T?. This results in unnecessary null checks at the usage sites, and
therefore the better way is to use the provided type exactly as it came
in, and if nullability is desired, then that must be explicitly
specified in the type parameter.