Learning pt 1: Naive attempt at using GridState in GridSection
This causes a crash in which the tile drag target no longer has a drop
target to snap to when the tile is released. The root cause is that
the original drop target is getting disposed in DragContext due to the
`onDragTargetAdded` and `onDragTargetRemoved` lambda references changing
(see the usage of `DisposableEffect` in `rememberDropTargetState`).

It turns out that the root cause of that happening is that Compose sees
that `onTilePlaced` and `onTileRemoved` make a reference to `gridState`,
which Compose has marked as 'unstable'. My understanding is that
composables with an unstable parameter will always recomposed, and it
seems that in this case, the `onTilePlaced/Removed` parameters also
became unstable because of their reference to `gridState`, hence
`GridCell`'s subsequent recomposition leading to the issue mentioned
above.

See https://developer.android.com/jetpack/compose/performance/stability

Resolution to this issue will come in a future commit.