Learning pt 2: Prevent crash from happening in GridSection
To fix the issue of the drop target state getting disposed, we can
simply remove the reference to the unstable `gridState` argument
from the `onTilePlaced` and `onTileRemoved` lambdas so that _those_
lambdas are no longer unstable (and GridCell won't get recomposed).
To do this, we just need to create a separate lambda that encapsulates
the `gridState` reference and pass that into `onTilePlaced/Removed`.

The following commit will address a discovered performance issue.