Add GridViewModel and GridState
I decided to keep grid-specific state and logic separate from the rest
of the game state (i.e. WordGameViewModel and WordGameState) because
those keep track of meta-information such as scores and available tiles,
which is not directly related to the state of the current board. Having
them be different classes ensures a clean separation of concerns.