Only recompose the exact GridCell whose state changed
I noticed that whenever a tile was placed, the entire loop inside of
`GridSection` was getting re-executed. Functionally, it was behaving
correctly as only the `GridCell` whose state was updated actually got
recomposed, but it is wasteful to loop through the entire grid only to
update a single cell.

The root cause turned out to be that the tile state (returned from
`onGetTile`) was being observed in `GridSection`, which led to
`GridSection` being the recomposition root instead of `GridCell`. By
pushing the invokation of `onGetTile` down into `GridCell`, we ensure
that Compose only recomposes the exact `GridCell` whose state was
updated.